# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/contacts/{_id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Find contact by ID.",
        "description": "This endpoint returns a contact found with the provided ID",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id of the desired contact.",
            "schema": {
              "type": "string",
              "example": "636c889a2a02ef8e6e9f50e6"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update contact.",
        "description": "This endpoint   a contact. All fields are required for a successful update.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phoneNumber": {
                  "example": "any"
                },
                "dateOfBirth": {
                  "example": "any"
                },
                "favoriteColor": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete contact.",
        "description": "This endpoint   a contact if a valid ID is passed.",
        "operationId": "Your_operationId_here",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
...
}